+++ src/Client.php
@@ -25,7 +25,7 @@ class Client {
      * @const string Current version of this client.
      * This follows Semantic Versioning (http://semver.org/)
      */
-    const VERSION = '1.6.2';
+    const VERSION = '1.7.0';
 
     /**
      * @const string The API endpoint for Notify production.
@@ -208,14 +208,37 @@ class Client {
      * @return array
      */
     public function sendLetter( $templateId, array $personalisation = array(), $reference = '' ){
-        
+
         $payload = $this->buildPayload( 'letter', '', $templateId, $personalisation, $reference );
 
         return $this->httpPost(
             self::PATH_NOTIFICATION_SEND_LETTER,
             $payload
         );
-        
+
+    }
+
+    /**
+     * Send a precompiled letter.
+     * Example usage: sendPrecompiledLetter($templateId, $ref, file_get_contents(<PATH TO FILE>)))
+     *
+     * @param string    $templateId
+     * @param string    $reference
+     * @param string    $pdf_data
+     *
+     * @return array
+     */
+    public function sendPrecompiledLetter( $reference, $pdf_data ){
+        $payload = [
+          'reference' => $reference,
+          'content' => base64_encode($pdf_data)
+        ];
+
+        return $this->httpPost(
+            self::PATH_NOTIFICATION_SEND_LETTER,
+            $payload
+        );
+
     }
 
     /**
@@ -257,7 +280,7 @@ class Client {
             'status',
             'template_type',
         ]));
-                
+
         return $this->httpGet( self::PATH_NOTIFICATION_LIST, $filters );
 
     }
@@ -336,6 +359,19 @@ class Client {
           'personalisation'=>$personalisation
         ];
         return $this->httpPost( $path, $payload );
+    }
+
+    /**
+     * Prepare a file before adding it to the $personalisation array for the sendEmail function
+     *
+     * @param string $file_contents
+     *
+     * @return string
+     */
+    public function prepareUpload( $file_contents ){
+      return [
+        "file" => base64_encode($file_contents)
+      ];
     }
 
     //------------------------------------------------------------------------------------